feat: migrate to signal and control flow
fix: workspace expanded signal

+++ packages/integration/src/lib/draw/draw.state.ts
@@ -30,8 +30,10 @@ export class DrawState {
           const deletedStore = this.stores.find(
             (store) => store.layer.id === layerId
           );
-          deletedStore.deleteMany(deletedStore.all());
-          this.stores.splice(this.stores.indexOf(deletedStore, 0), 1);
+          if (deletedStore) {
+            deletedStore.deleteMany(deletedStore.all());
+            this.stores.splice(this.stores.indexOf(deletedStore, 0), 1);
+          }
           this.layersID.splice(this.layersID.indexOf(layerId, 0), 1);
           const drawControlIndex = this.drawControls.findIndex(
             (dc) => dc[0] === layerId
